+++ packages/fpdart/example/src/typeclass/eq/eq1.dart
@@ -8,12 +8,12 @@ class Parent {
 
 void main() {
   /// Equality for values of type [Parent] based on their `value1` ([int]).
-  final eqParentInt = Eq.eqInt().contramap<Parent>(
+  final eqParentInt = Eq.eqInt.contramap<Parent>(
     (p) => p.value1,
   );
 
   /// Equality for of type [Parent] based on their `value2` ([double]).
-  final eqParentDouble = Eq.eqDouble().contramap<Parent>(
+  final eqParentDouble = Eq.eqDouble.contramap<Parent>(
     (p) => p.value2,
   );
 }